% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BipartiteGroupIPW_function.R
\name{BipartiteGroupIPW}
\alias{BipartiteGroupIPW}
\title{Function that estimates the cluster average potential outcome in settings of
bipartite interference.}
\usage{
BipartiteGroupIPW(int_dta, out_dta, cov_cols, phi_hat, alpha,
  trt_col = NULL, out_col = NULL, integral_bound = 10)
}
\arguments{
\item{int_dta}{Data on the interventional units including treatment,
neighborhood index and covariates. Neighborhood variable should be named
'neigh' and only include continuous integers from 1 to the number of
neighborhoods.}

\item{out_dta}{Data on the outcome units including outcome, neighborhood,
and index of the closest interventional unit as 'closest_int'.
Neighborhood variable should be named 'neigh' and only include continuous
integers from 1 to the number of neighborhoods.}

\item{cov_cols}{The indices including the covariates of the ps model.}

\item{phi_hat}{A list with two elements. The first one is a vector of
coefficients of the ps, and the second one is the random effect variance.}

\item{alpha}{The values of alpha for which we want to estimate the group
average potential outcome.}

\item{trt_col}{If the treatment is not named 'A' in int_dta, specify the
treatment column index.}

\item{out_col}{If the outcome is not named 'Y' in out_dta, specify the
outcome column index.}

\item{integral_bound}{The number of standard deviations of the random effect
that will be used as the lower and upper limit. Defaults to 10.}
}
\description{
Function that estimates the cluster average potential outcome in settings of
bipartite interference.
}
