% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootVar_function.R
\name{BootVar}
\alias{BootVar}
\title{Bootstrap variance of potential outcomes.}
\usage{
BootVar(dta, B = 500, alpha, ps = c("true", "est"), cov_cols,
  phi_hat_true = NULL, ps_info_est = NULL, verbose = TRUE,
  ps_specs = NULL, trt_col = NULL, out_col = NULL, return_var = FALSE)
}
\arguments{
\item{dta}{The data frame including the observed data set.}

\item{B}{Number of bootstrap samples.}

\item{alpha}{The values of alpha where the potential outcomes are estimated.}

\item{ps}{Character. Whether the propensity score is known or is estimated.
Options include 'true', 'est'. Defaults to 'true'.}

\item{cov_cols}{Vector of column indices of the covariates used in the
propensity score model.}

\item{phi_hat_true}{List. Specify if the propensity score is known (ps set
to 'true'). Elements of the list are trt_coef and re_var including the
coefficients of the propensity score and random effect variance.}

\item{ps_info_est}{List of elements for acquiring estimates based on the
estimated propensity score. The list includes 1) glm_form: Element of 
formula class. The formula can be either for a fixed effects model or for a
model including random intercepts. 2) ps_with_re: An indicator of whether
the propensity score is a mixed model (set to TRUE) or not (set to FALSE).
3) gamma_numer: The coefficients of the covariates in the counterfactual
treatment allocation model, and 4) use_control: Set to TRUE or FALSE if
you want or do not want additional elements in fitting the mixed model.
use_control does not have to be specified.}

\item{verbose}{Logical. Whether progress is printed. Defaults to TRUE.}

\item{trt_col}{If the treatment is not named 'A' in dta, specify the
treatment column index.}

\item{out_col}{If the outcome is not named 'Y', specify the outcome column
index.}

\item{return_var}{Logical. Defaults to FALSE. Whether the function should
return indicators of whether the random effect variance is estimated to be
zero. If ps is set to true, or ps_info_est sets ps_with_re equal to FALSE,
this argument is necessarily FALSE.}
}
\description{
Using re-sampling of clusters to acquire an estimate of the potential
outcome estimator variance.
}
