% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcScore_function.R
\name{CalcScore}
\alias{CalcScore}
\title{Matrix of score for the propensity score.}
\usage{
CalcScore(
  dta,
  neigh_ind = NULL,
  phi_hat,
  cov_cols,
  trt_name = NULL,
  integral_bound = 10
)
}
\arguments{
\item{dta}{The data set including (at least) the treatment and covaratiates.}

\item{neigh_ind}{A list including one element for each neighborhood. That
element includes the indices of the observations in dta that belong to each
neighborhood. Can be left NULL if dta includes a column neigh.}

\item{phi_hat}{A list with two elements. The first one is a vector of
coefficients of the propensity score, and the second one is the random
effect variance.}

\item{cov_cols}{The indeces including the covariates of the propensity score
model.}

\item{trt_name}{The name of the treatment column. If it is 'A', you can
leave NULL.}
}
\description{
Matrix of score for the propensity score.
}
