% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_function.R
\name{DE}
\alias{DE}
\title{Direct effect estimates and asymptotic variance.}
\usage{
DE(ypop, ypop_var, alpha = NULL)
}
\arguments{
\item{ypop}{A matrix with rows corresponding to the potential outcome under
control and treatment, and columns corresponding to the cluster-average
propensity of treatment.}

\item{ypop_var}{An 3-dimensional array, where the first two dimensions are
equal to 2 and include the variance covariance matrix of the population
average potential outcome for each alpha. Dimension 3 is alpha.}

\item{alpha}{The values of alpha we consider. If ypop has column names,
alpha can be left null.}
}
\value{
A matrix with rows including the estimate and variance of the direct
effect and columns corresponding to alpha.
}
\description{
Direct effect estimates and asymptotic variance.
}
