% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Denominator_function.R
\name{Denominator}
\alias{Denominator}
\title{Calculating the denominator of the group estimator.}
\usage{
Denominator(A, X, phi_hat, alpha = NULL, integral_bound = 10,
  include_alpha = TRUE)
}
\arguments{
\item{A}{The treatment vector for units in the group.}

\item{X}{The covariates for units in the group.}

\item{phi_hat}{A list with two elements. The first one is a vector of
coefficients of the ps, and the second one is the random effect variance.
If the second element is 0, the propensity score excludes random effects.}

\item{alpha}{This value of alpha is used to stabilize calculations. If
include_alpha is set to TRUE, alpha needs to coincide with the alpha used
for calculating the numerator.}

\item{integral_bound}{If the propensity score includes a random effect, the
integral in the denominator is calculated over the normal distribution of
the random effects from - integral_bound to integral_bound.}

\item{include_alpha}{If include_alpha is set to true, the probabilities in
the denominator are divided by the specified value of alpha to stabilize the
integral calculation.}
}
\description{
Calculating the denominator of the group estimator.
}
