% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DenomIntegral_function.R
\name{DenomIntegral}
\alias{DenomIntegral}
\title{Function that calculates the integral in the denominator corresponding to the
marginal likelihood of the vector A, given X and the fixed effects.}
\usage{
DenomIntegral(A, X, phi_hat, alpha = NULL, integral_bound = 10,
  include_alpha = TRUE)
}
\description{
Function that calculates the integral in the denominator corresponding to the
marginal likelihood of the vector A, given X and the fixed effects.
}
