% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{estimates}
\alias{estimates}
\alias{estimators_full_neighborhood}
\title{Estimate exposure-specific causal effects.}
\usage{
estimates(
  obs_exposure,
  obs_outcome,
  obs_prob_exposure,
  n_var_permutations = 10,
  effect_estimators = c("hajek", "horvitz-thompson"),
  variance_estimators = c("hajek", "horvitz-thompson"),
  hop
)

estimators_full_neighborhood(
  obs_exposure,
  obs_outcome,
  obs_prob_exposure,
  n_var_permutations = 10
)
}
\arguments{
\item{obs_exposure}{an \code{N}\eqn{*} K named numeric matrix of indicators for
whether units \code{N} are in exposure condition k, where K is the total number
of exposure conditions and names correspond to the exposure conditions.
Such matrix is returned by function \code{\link{make_exposure_map_AS}}.}

\item{obs_outcome}{a vector length \code{N} of outcome data.}

\item{obs_prob_exposure}{a list of 3 lists containing exposure probabilities:
\describe{ \item{\code{I_exposure}:}{A list of K \code{N} \eqn{*} \code{R} numeric
matrices of indicators for whether units \code{N} are in exposure condition k
over each of the possible \code{R} treatment assignment vectors. The number of
numeric matrices K corresponds to the number of exposure conditions.}
\item{\code{prob_exposure_k_k}:}{A list of K symmetric \code{N} \eqn{*} \code{N} numeric
matrices each containing individual exposure probabilities to condition k
on the diagonal, and joint exposure probabilities to condition k on the
off-diagonals.} \item{\code{prob_exposure_k_l}:}{A list of
\eqn{permutation(K,2)} nonsymmetric \code{N} \eqn{*} \code{N} numeric matrices each
containing joint probabilities across exposure conditions k and l on the
off-diagonal, and zeroes on the diagonal. When K = 4, the number of numeric
matrices is 12; \eqn{permutation(4,2)}.} } Such list is returned by
function \code{\link{make_exposure_prob}}.}

\item{n_var_permutations}{number of tratment permutations to estimate the
constant effects variance estimator derived in Aronow (2013). Default is
\code{10}, but must be smaller or equal to \code{R}, the number of permutations to
compute exposure probabilities. Recommended is \code{1000}, when \code{R} \eqn{>
  1000}.}

\item{effect_estimators}{string vector with names of estimators to be estimated
among 'hajek', 'horvitz-thompson'. Default is both.}

\item{variance_estimators}{string vector with names of variance estimators
to be estimated among 'hajek', 'horvitz-thompson', 'constant_effect',
'max_ht_const'. Default includes the first two. Estimating 'constant_effect'
or 'max_ht_const' signficantly increases the running time.}

\item{hop}{number; either \code{1} or \code{2}. Must be \code{1} if argument \code{hop = 1} in
function \code{\link{make_exposure_map_AS}} which assumes first-degree
interference and produces four exposure conditions. Must be \code{2} if argument
\code{hop = 2} in function \code{\link{make_exposure_map_AS}} which assumes
second-degree interference and produces eight exposure conditions.}
}
\value{
A list of 13 lists: \describe{ \item{\code{yT_ht}:}{A named numeric vector
which contains the values of the Horvitz-Thompson estimator of the total of
potential outcomes under each exposure condition as derived in Equation 1
of Aronow and Samii (2017).} \item{\code{yT_h}:}{A named numeric vector which
contains the values of the Hajek estimator of the total of potential
outcomes under each exposure condition as derived in Equation 15 of Aronow
and Samii (2017).} \item{\code{var_yT_ht}:}{A named numeric K \eqn{*} 1 matrix
which contains the values of the variance estimator of the Horvitz-Thompson
estimator of the total of potential outcomes under each exposure condition
as derived in Equation 7 and Proposition 5.1 of Aronow and Samii (2017).}
\item{\code{var_yT_h}:}{A named numeric K \eqn{*} 1 matrix which contains the
values of the variance estimator of the Hajek estimator of the total of
potential outcomes under each exposure condition as explained in the first
paragraph of page 1929 of Aronow and Samii (2017).} \item{\code{cov_yT_ht}:}{A
named numeric \eqn{permutation(K,2)} \eqn{*} 1 matrix which contains the
values of the covariance estimator of the Horvitz-Thompson estimator of the
total of potential outcomes across exposures conditions k and l as derived
in Equation 10 of Aronow and Samii (2017). When the number of exposure
conditions K = 4, then the number of rows of this matrix is 12;
\eqn{permutation(4,2)}.} \item{\code{cov_yT_h}:}{A named numeric
\eqn{permutation(K,2)} \eqn{*} 1 matrix which contains the values of the
covariance estimator of the Hajek estimator of the total of potential
outcomes across exposures conditions k and l as explained in the first
paragraph of page 1929 of Aronow and Samii (2017). When the number of
exposure conditions K = 4, then the number of rows of this matrix is 12;
\eqn{permutation(4,2)}.} \item{\code{tau_ht}:}{A named numeric vector which
contains the  values of the Horvitz-Thompson estimator of the average
unit-level causal effect of exposure k versus exposure l as derived in
Equation 3 of Aronow and Samii (2017). Here exposure l is fixed to the
\eqn{No Exposure} condition (i.e. no direct or indirect exposure).}
\item{\code{tau_h}:}{A named numeric vector which contains the values of the
Hajek estimator of the average unit-level causal effect of exposure k
versus exposure l. Here exposure l is fixed to the \eqn{No Exposure}
condition (i.e. no direct or indirect exposure).} \item{\code{tau_dsm}:}{A named
numeric vector which contains the values of the difference in sample means
estimator of the total observed outcomes across exposures k and l. Here
exposure l is fixed to the \eqn{No Exposure} condition (i.e. no direct or
indirect exposure).} \item{\code{var_tau_ht}:}{A named numeric vector which
contains the values of the conservative variance estimator of the variance
of the Horvitz-Thompson estimator of the average unit-level causal effect
of exposure k versus exposure l as derived in Equation 11 of Aronow and
Samii (2017). Here exposure l is fixed to the \eqn{No Exposure} condition
(i.e. no direct or indirect exposure).} \item{\code{var_tau_h}:}{A named numeric
vector which contains the values of the linearized variance estimator of
the variance of the Hajek estimator of the average unit-level causal effect
of exposure k versus exposure l as derived in Equation 11 of Aronow and
Samii (2017) and further explained in the first paragraph of page 1929.
Here exposure l is fixed to the \eqn{No Exposure} condition (i.e. no direct
or indirect exposure).} \item{\code{var_tau_ht_const_eff}:}{A named numeric
vector which contains the values of the constant effects variance
estimator of the variance of the Horvitz-Thompson estimator of the average
unit-level causal effect of exposure k versus exposure l as derived in
Equation 2.15 of Aronow (2013). Here exposure l is fixed to the \eqn{No
  Exposure} condition (i.e. no direct or indirect exposure).}
\item{\code{var_tau_ht_max}:}{A named numeric vector which contains the maximum
between \code{var_tau_h} and \code{var_tau_ht_const_eff}.} }
}
\description{
Estimate exposure-specific causal effects and their variance.
}
\details{
\code{estimates} produces values for the estimator of the average unit-level
causal effect of exposure k versus l and its variance estimator for the
exposure mappings returned by function \code{\link{make_exposure_map_AS}},
using a Horvitz-Thompson and a Hajek estimator. It also computes
Horvitz-Thompson and Hajek estimators of the total of potential outcomes,
which are inputs in the computation of average unit-level causal effect of
exposure k versus l.
}
\section{Functions}{
\itemize{
\item \code{estimators_full_neighborhood}: Produces values for the estimator of the average
unit-level causal effect of exposure k versus l and its variance estimator
for the exposure mapping returned by function
\code{\link{make_exposure_map_full_neighborhood}}, using a Horvitz-Thompson
and a Hajek estimator. It also computes Horvitz-Thompson and Hajek
estimators of the total of potential outcomes, which are inputs in the
computation of average unit-level causal effect of exposure k versus l.
}}

\examples{
# Create adjacency matrix and treatment assignment vector
# to produce observed exposure conditions:

adj_matrix <- make_adj_matrix(N = 9, model = 'sq_lattice')

tr_vector <- make_tr_vec_permutation(N = 9, p = 0.2,
                                     R = 1, seed = 357)

obs_exposure <- make_exposure_map_AS(adj_matrix, tr_vector,
                                     hop = 1)

# Simulate a vector of outcome data:

potential_outcome <- make_dilated_out(adj_matrix, make_corr_out,
                                      seed = 357, hop = 1)

obs_outcome <- rowSums(obs_exposure*t(potential_outcome))

# Create exposure probabilities:

potential_tr_vector <- make_tr_vec_permutation(N = 9, p = 0.2,
                                               R = 36,
                                               seed = 357)

obs_prob_exposure <- make_exposure_prob(potential_tr_vector,
                                        adj_matrix,
                                        make_exposure_map_AS,
                                        list(hop=1))

Estimate exposure-specific causal effects and their variance:

estimates(obs_exposure, obs_outcome, obs_prob_exposure,
                                     n_var_permutations = 30,
                                     hop = 1)
# Create adjacency matrix and treatment vector to
# produce observed exposure conditions according to the
# "full neighborhood" exposure mapping:

adj_matrix <- make_adj_matrix(N = 81, model = 'sq_lattice')

tr_vector <- make_tr_vec_permutation(N = 81, p = 0.5,
                                     R = 1, seed = 357)

obs_exposure_full_nei <- make_exposure_map_full_neighborhood(adj_matrix,
                                                    tr_vector)
# Simulate a vector of outcome data:

potential_outcome_full_nei <-
  make_dilated_out_full_neighborhood(adj_matrix, make_corr_out,
                                     seed = 357)

obs_outcome_full_nei <-
  rowSums(obs_exposure_full_nei*t(potential_outcome_full_nei))

# Create exposure probabilities:

potential_tr_vector <- make_tr_vec_permutation(N = 81, p = 0.5,
                                               R = 36,
                                               seed = 357)
obs_prob_exposure_full_nei <- make_exposure_prob(potential_tr_vector,
                                        adj_matrix,
                                        make_exposure_map_full_neighborhood)

Estimate exposure-specific causal effects and their variance:

estimators_full_neighborhood(obs_exposure_full_nei, obs_outcome_full_nei,
          obs_prob_exposure_full_nei,
          n_var_permutations = 30)
}
\references{
Aronow, P. M. (2013). \href{https://search.proquest.com/docview/1567045106?accountid=12768}{Model assisted causal inference}.
\emph{PhD thesis, Department of Political Science, Yale University, New Haven,
CT}.

Aronow, P.M. & Samii, C. (2017). \href{https://doi.org/10.1214/16-AOAS1005}{Estimating average causal effects under general interference, with application to a social network experiment}. \emph{The Annals of Applied
Statistics}, 11(4), 1912--1947.

Aronow, P.M. et al. (2020). \href{https://arxiv.org/abs/2001.05444}{Spillover effects in experimental data}. \emph{arXiv preprint},
arXiv:2001.05444.
}
