% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates_hierarchical.R
\name{estimates_hierarchical}
\alias{estimates_hierarchical}
\title{Estimate average causal effects under a two-stage hierarchical design.}
\usage{
estimates_hierarchical(estimator_data)
}
\arguments{
\item{estimator_data}{data frame of \code{N} observations and the variables:
\describe{ \item{\code{group}:}{Integer vector specifying the group label.}
\item{\code{group_tr}:}{Numeric indicator of group assignment to high saturation
(psi) in the first stage.} \item{\code{indiv_tr}:}{Numeric indicator of
individual assignment to treatment in the second stage.}
\item{\code{obs_outcome}:}{Numeric vector of observed outcomes.} }}
}
\value{
A list of 2 lists: \enumerate{ \item A list of 5 scalars corresponding
to the values of the estimators of the average direct causal effect under
high treatment saturation (psi), the average direct causal effect under low
treatment saturation (phi), the average indirect causal effect, the average
total causal effect, and the average overall causal effect. \item A list of
5 scalars corresponding to the values of the variance of the estimator of
the average direct causal effect under high treatment saturation (psi), the
average direct causal effect under low treatment saturation (phi), the
average indirect causal effect, the average total causal effect, and the
average overall causal effect. }
}
\description{
Estimate average causal effects and their variance under two-stage
hierarchical treatment assignment, assuming partial and stratified
interference.
}
\details{
\code{estimates} produces values of the estimators proposed by Hudgens and
Halloran (2008) of the population average direct causal effect under high
treatment saturation (psi), the population average direct causal effect under
low treatment saturation (phi), the population average indirect causal effect,
the population average total causal effect, and the population average overall
causal effect. It also produces values of the variance estimators which assume
stratified interference (i.e. potential outcomes of a unit are affected by its
own treatment assignment and only the treated proportion of its group; the
precise set of treated group members does not matter).
}
\examples{
# Simulate first and second stage treatment
# assignment and outcomes:

group <- rep(1:6, each = 30/6)
c <- 1/2
k <- c(2/5, 3/5)

tr_assignment <-
make_tr_vec_permutation_hierarchical(group, c, k, R = 1,
                                     seed = 357)[[1]]

potential_outcomes <- make_dilated_out_hh(N = 30,
                                          seed = 357)

# Create data frame with group label, first and second stage
# treatment assignment and outcomes:

estimator_data <- make_estimator_data(tr_assignment,
                                      potential_outcomes)

# Estimate average causal effects and their variance:
estimates_hierarchical(estimator_data)
}
\references{
Hudgens, M.G. & Halloran M.E. (2008). \href{https://doi.org/10.1198/016214508000000292}{Toward causal inference with interference}. \emph{Journal of
the American Statistical Association}, 103(482), 832--842.

Aronow, P.M. et al. (2020). \href{https://arxiv.org/abs/2001.05444}{Spillover effects in experimental data}. \emph{arXiv preprint}, arXiv:2001.05444.
}
