% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FromAlphaToRE_function.R
\name{FromAlphaToRE}
\alias{FromAlphaToRE}
\title{Function that calculates the random effect that corresponds to a specific
average probability of treatment.}
\usage{
FromAlphaToRE(alpha, lin_pred, alpha_re_bound = 10)
}
\arguments{
\item{alpha}{Target average probability of treatment. Numberic in 0 - 1.}

\item{lin_pred}{Linear predictor for the probability of treatment among the
remaining units. Includes intercept and fixed effects.}

\item{alpha_re_bound}{The lower and upper end of the values for bi we will
look at. Defaults to 10, meaning we will look between - 10 and 10.}
}
\description{
Function that calculates the random effect that corresponds to a specific
average probability of treatment.
}
