% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSimData_function.R
\name{GetSimData}
\alias{GetSimData}
\title{Simulating treatment assignment for interference data.}
\usage{
GetSimData(dta, pot_out, neigh_ind, re_sd, trt_coef)
}
\arguments{
\item{dta}{A data frame including the following variables: neigh varying from 1 to
the number of neighborhoods, and two covariates X1 and X2.}

\item{pot_out}{A list of length equal to the number of clusters. Each element is an
array of 3 dimensions. Element ijk of the array is the potential outcome for unit i
in the cluster if the treatment of unit is j - 1 and k - 1 of neighbors are treated.}

\item{neigh_ind}{A list of length equal to the number of custers including the row
indices corresponding to each cluster.}

\item{re_sd}{The standard deviation of the cluster specific random effect for the
propensity score model.}

\item{trt_coef}{Coefficients of the propensity score model corresponding to the
intercept, and the two covariates in dta.}
}
\description{
Simulating treatment assignment for interference data.
}
