% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IE_function.R
\name{IE}
\alias{IE}
\title{Indirect effect estimates and asymptotic variance.}
\usage{
IE(ygroup, ps = c("true", "estimated"), scores = NULL)
}
\arguments{
\item{ypop}{A vector of the population average potential outcome for the
values of alpha.}

\item{ypop_var}{An array with dimensions 2, 2, alpha, alpha, including the
covariance matrix of the average potential outcome under two specifications
of alpha.}

\item{alpha}{The values of alpha we consider.}
}
\value{
A 3-dimensional array with dimension 1 equal to 2 including the
estimate and the asymptotic variance. The remaining two dimensions
correspond to values of alpha.
}
\description{
Indirect effect estimates and asymptotic variance.
}
