% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IEvar_function.R
\name{IEvar}
\alias{IEvar}
\title{Asymptotic variance of population average potential outcome vector for
different alphas.}
\usage{
IEvar(ygroup, ps = c("true", "estimated"), scores = NULL)
}
\arguments{
\item{ygroup}{An matrix including the group average potential outcome
estimates where rows correspond to group, and columns to values of alpha.}

\item{ps}{String. Can take values 'true', or 'estimated' for known or
estimated propensity score. Defaults to 'true'.}

\item{scores}{A matrix with rows corresponding to the parameters of the
propensity score model and columns for groups. Includes the score of the
propensity score evaluated for the variables of each group. Can be left
NULL for ps = 'true'.}
}
\description{
Asymptotic variance of population average potential outcome vector for
different alphas.
}
