% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dilated_out.R
\name{make_dilated_out}
\alias{make_dilated_out}
\alias{make_dilated_out_full_neighborhood}
\title{Create potential dilated outcomes.}
\usage{
make_dilated_out(adj_matrix, make_corr_out, seed, hop, multipliers = NULL)

make_dilated_out_full_neighborhood(
  adj_matrix,
  make_corr_out,
  multipliers = NULL,
  seed = NULL
)
}
\arguments{
\item{adj_matrix}{an \code{N} \eqn{*} \code{N} numeric matrix of 0, 1 entries such as those
retuned by \code{\link{make_adj_matrix}}, or a \code{N} \eqn{*} \code{N} matrix, where \code{N}
is the number of units.}

\item{make_corr_out}{unused.}

\item{seed}{random number for result replicability.}

\item{hop}{number; either \code{1} or \code{2}. If \code{1} assumes first-degree
interference and produces four exposure conditions, if \code{2} assumes
second-degree interference and produces eight exposure conditions.}

\item{multipliers}{numeric vector with dilated effects multipliers. Must be
of length 3 if \code{hop} is 1, and length 7 if \code{hop} is 2. Default are
\code{c(2,1.5,1.25)} and \code{c(2.25,2,1.75,1.5,1.375,1.25,1.125)}.}
}
\value{
An K \eqn{*} \code{N} named numeric matrix, where K corresponds to the number of
exposure conditions and \code{N} number of units.
}
\description{
Create potential dilated outcomes for simulation.
}
\details{
\code{make_dilated_out} produces potential dilated outcomes for simulation
according to two different exposure mappings. An exposure mapping that
assumes that interference happens only through direct peer connections
(first-degree interference) which produces four exposure conditions:
\eqn{Direct + Indirect Exposure}, \eqn{Isolated Direct Exposure},
\eqn{Indirect Exposure}, \eqn{No Exposure}. And an exposure mapping that
assumes second-degree interference which produces eight exposure conditions.
The values for the baseline \eqn{No Exposure} condition are drawn from an
absolute standard normal distribution which is correlated with the unit's
first and second order degree. The values for the other exposure conditions
are obtained by multiplying the vector of multipliers by the baseline \eqn{No
Exposure} value.
}
\section{Functions}{
\itemize{
\item \code{make_dilated_out_full_neighborhood}: Produces potential dilated outcomes for simulation according to
an exposure mapping that assumes first-degree interference and that
produces two exposure conditions: \eqn{All Treat}, a condition in which a
unit and 100\% of its first-degree neighbors are treated, and \eqn{All Control},
a unit and 100\% of its first-degree neighbors are in the control
condition.
}}

\examples{
adj_matrix <- make_adj_matrix(N = 9, model = 'sq_lattice')
multipliers <- c(4, 2, 3)
make_dilated_out(adj_matrix, make_corr_out, seed = 357,
multipliers = multipliers, hop = 1)
make_dilated_out(adj_matrix, make_corr_out, seed = 357,
multipliers = multipliers, hop = 2)
multiplier <- 4
make_dilated_out_full_neighborhood(adj_matrix, make_corr_out,
                                   multipliers=multiplier,
                                   seed=357)    
}
\references{
Aronow, P.M. et al. (2020). \href{https://arxiv.org/abs/2001.05444}{Spillover effects in experimental data}. \emph{arXiv preprint},
arXiv:2001.05444.
}
