% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dilated_out_hh.R
\name{make_dilated_out_hh}
\alias{make_dilated_out_hh}
\title{Create potential dilated outcomes under a two-stage hierarchical design.}
\usage{
make_dilated_out_hh(
  N,
  out = function(x) abs(rnorm(x)),
  multipliers = NULL,
  seed = NULL
)
}
\arguments{
\item{N}{number of units.}

\item{out}{function which returns a vector length \code{N} of outcome values for
treatment condition \eqn{Indirect Phi} which is the baseline (or control)
condition. Default function draws from an absolute standard normal
distribution.}

\item{multipliers}{numeric vector with dilated effects multipliers. Must be
of length 3. Default is \code{c(2,1.5,1.25)}.}

\item{seed}{random number for result replicability.}
}
\value{
An K \eqn{*} \code{N} named numeric matrix, where K corresponds to the
number of exposure conditions and \code{N} number of units.
}
\description{
Create potential dilated outcomes for simulation under two-stage hierarchical
treatment assignment, assuming partial and stratified interference.
}
\details{
\code{make_dilated_out_hh} produces potential dilated outcomes for simulation
according to a two-stage hierarchical design where groups are first randomly
assigned to a high or a low level of treatment saturation (psi, phi), and
then units within groups are randomly assigned to treatment with probability
equal to their group saturation rate. Following Hudgens and Halloran (2008),
there are four potential outcomes which correspond to the four potential
treatment conditions: \eqn{Direct + Indirect Psi}--a unit is directly treated
and its group is assigned treatment saturation psi, \eqn{Direct + Indirect
Phi}--a unit is directly treated and its group is assigned treatment
saturation phi, \eqn{Indirect Psi}--a unit is not directly treated and its
group is assigned treatment saturation psi, \eqn{Indirect Phi}--a unit is not
directly treated and its group is assigned treatment saturation phi. This
function assumes stratified interference (i.e. potential outcomes of a unit
are affected by its own treatment assignment and only the treated proportion
of its group; the precise set of treated group members does not matter).
}
\examples{
make_dilated_out_hh(N = 10, seed = 357)

multipliers <- c(4, 3, 2)
make_dilated_out_hh(N = 10, multipliers = multipliers, seed = 357)
}
\references{
Hudgens, M.G. & Halloran M.E. (2008). \href{https://doi.org/10.1198/016214508000000292}{Toward causal inference with interference}. \emph{Journal of
the American Statistical Association}, 103(482), 832--842.

Aronow, P.M. et al. (2020). \href{https://arxiv.org/abs/2001.05444}{Spillover effects in experimental data}. \emph{arXiv preprint},
arXiv:2001.05444.
}
