% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_exposure_map_AS.R
\name{make_exposure_map_AS}
\alias{make_exposure_map_AS}
\alias{make_exposure_map_full_neighborhood}
\title{Create treatment exposure conditions.}
\usage{
make_exposure_map_AS(adj_matrix, tr_vector, hop)

make_exposure_map_full_neighborhood(adj_matrix, tr_vector)
}
\arguments{
\item{adj_matrix}{an \code{N} \eqn{*} \code{N} numeric matrix of 0, 1 entries such as those
retuned by \code{\link{make_adj_matrix}}, or a \code{N} \eqn{*} \code{N} matrix, where \code{N}
is the number of units.}

\item{tr_vector}{a numeric vector of length \code{N} of 0, 1 entries such as that
retuned by \code{\link{make_tr_vec_permutation}} with \code{R = 1}, or a vector
of length \code{N} with the treatment assignment.}

\item{hop}{number; either \code{1} or \code{2}. If \code{1} assumes first-degree
interference and produces four exposure conditions, if \code{2} assumes
second-degree interference and produces eight exposure conditions.}
}
\value{
An \code{N} \eqn{*} K numeric matrix, where K corresponds to the number of
exposure conditions.
}
\description{
Create treatment exposure conditions according to selected exposure mappings.
}
\details{
\code{make_exposure_map_AS} produces treatment exposure conditions according
to two different exposure mappings. An exposure mapping that assumes that
interference happens only through direct peer connections (first-degree
interference) which produces four exposure conditions: \eqn{Direct + Indirect
Exposure}, \eqn{Isolated Direct Exposure}, \eqn{Indirect Exposure}, \eqn{No
Exposure}. And an exposure mapping that assumes second-degree interference
which produces eight exposure conditions.
}
\section{Functions}{
\itemize{
\item \code{make_exposure_map_full_neighborhood}: Create treatment exposure conditions
according to an exposure mapping that assumes first-degree interference and
that produces two exposure conditions: \eqn{All Treat}; a unit and 100\% of
its first-degree neighbors are treated, and \eqn{All Control}; a unit and
100\% of its first-degree neighbors are in the control condition.
}}

\examples{
# Create an adjacency matrix and a treatment vector:

adj_matrix <- make_adj_matrix(N = 9, model = 'sq_lattice')
tr_vector <- make_tr_vec_permutation(N = 9, p = 0.2, R = 1,
                                     seed = 357)

# Create treatment exposure conditions:

make_exposure_map_AS(adj_matrix, tr_vector, hop = 1)
make_exposure_map_AS(adj_matrix, tr_vector, hop = 2)
make_exposure_map_full_neighborhood(adj_matrix, tr_vector)
}
\references{
Aronow, P.M. & Samii, C. (2017). \href{https://doi.org/10.1214/16-AOAS1005}{Estimating average causal effects under general interference, with application to a social network experiment}. \emph{The Annals of Applied
Statistics}, 11(4), 1912--1947.

Aronow, P.M. et al. (2020). \href{https://arxiv.org/abs/2001.05444}{Spillover effects in experimental data}. \emph{arXiv preprint},
arXiv:2001.05444.
}
