% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_exposure_prob.R
\name{make_exposure_prob}
\alias{make_exposure_prob}
\title{Create exposure probabilities.}
\usage{
make_exposure_prob(
  potential_tr_vector,
  adj_matrix,
  exposure_map_fn,
  exposure_map_fn_add_args = NULL
)
}
\arguments{
\item{potential_tr_vector}{an \code{R} \eqn{*}  \code{N} matrix of 0, 1 entries such as that
produced by \code{\link{make_tr_vec_permutation}}, or an \code{R} \eqn{*}  \code{N} matrix
containing \code{R} permuted treatment assignments for \code{N} units.}

\item{adj_matrix}{an \code{N} \eqn{*} \code{N} numeric matrix of 0, 1 entries such as those
retuned by \code{\link{make_adj_matrix}}, or a \code{N} \eqn{*} \code{N} matrix, where \code{N}
is the number of units.}

\item{exposure_map_fn}{function which returns the exposure mapping such as
\code{\link{make_exposure_map_AS}}.}

\item{exposure_map_fn_add_args}{list of additional arguments which are passed
to \code{exposure_map_fn}. \code{adj_matrix} and \code{tr_vector} should not be specified
here because they are already passed with values derived from the arguments
to \code{make_exposure_prob}. If using \code{\link{make_exposure_map_AS}}, \code{hop}
must be specified here.}
}
\value{
A list of 3 lists: \describe{ \item{\code{I_exposure}:}{A list of K
\code{N} \eqn{*}  \code{R} numeric matrices of indicators for whether units \code{N} are in
exposure condition k over each of the possible \code{R} treatment assignment
vectors. The number of numeric matrices K corresponds to the number of
exposure conditions.} \item{\code{prob_exposure_k_k}:}{A list of K symmetric
\code{N} \eqn{*}  \code{N} numeric matrices each containing individual exposure probabilities
to condition k on the diagonal, and joint exposure probabilities to
condition k on the off-diagonals.} \item{\code{prob_exposure_k_l}:}{A list of
\eqn{permutation(K,2)} nonsymmetric \code{N}\eqn{*} \code{N} numeric matrices each
containing joint probabilities across exposure conditions k and l on the
off-diagonal, and zeroes on the diagonal. When K = 4, the number of numeric
matrices is 12; \eqn{permutation(4,2)}.}}
}
\description{
Create matrices of exposure probabilities.
}
\details{
\code{make_exposure_prob} produces the units' probabilities of being subject
to each of the possible exposure conditions which are used for estimating
exposure-specific causal effects, and the joint exposure probabilities used
for variance estimators.
}
\examples{
potential_tr_vector <- make_tr_vec_permutation(N = 9, p = 0.2, R = 20, seed = 357)
adj_matrix <- make_adj_matrix(N = 9, model = 'sq_lattice')
make_exposure_prob(potential_tr_vector, adj_matrix, make_exposure_map_AS, list(hop=1))
}
\references{
Aronow, P.M. & Samii, C. (2017). \href{https://doi.org/10.1214/16-AOAS1005}{Estimating average causal effects under general interference, with application to a social network experiment}. \emph{The Annals of Applied
Statistics}, 11(4), 1912--1947.

Aronow, P.M. et al. (2020). \href{https://arxiv.org/abs/2001.05444}{Spillover effects in experimental data}. \emph{arXiv preprint},
arXiv:2001.05444.
}
