% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tr_vec_bernoulli.R
\name{make_tr_vec_bernoulli}
\alias{make_tr_vec_bernoulli}
\title{Assign treatment via 3-net clustering randomization.}
\usage{
make_tr_vec_bernoulli(adj_matrix, p, R, cluster, seed = NULL)
}
\arguments{
\item{adj_matrix}{an \code{N} \eqn{*} \code{N} numeric matrix of 0, 1 entries such as those
retuned by \code{\link{make_adj_matrix}}, or a \code{N} \eqn{*} \code{N} matrix, where \code{N}
is the number of units.}

\item{p}{probability that treatment takes the value 1.}

\item{R}{number of repetitions (treatment assignments). \code{R} must be smaller
or equal to the number of possible treatment assignements which is
\eqn{combination(N, C) * 2 ^ C} when \code{cluster = 'yes'}, and \eqn{2 ^
  N} when \code{cluster = 'no'}, where N corresponds to the number of units
and C to the number of clusters.}

\item{cluster}{string; either \code{'yes'} or \code{'no'}. If \code{'yes'} units are
assigned to treatment via 3-net clustering randomization, if \code{'no'} via
unit-level randomization under a Bernoulli(p) distribution.}

\item{seed}{random number for result replicability.}
}
\value{
An \code{R} \eqn{*} \code{N} numeric matrix. Each row cooresponds to a
treatment assignment vector.
}
\description{
Create a permutation matrix from possible treatment assignments via 3-net
clustering randomization or via unit-level randomization under a Bernoulli
distribution.
}
\details{
\code{make_tr_vec_bernoulli} produces all possible (or a random sample of all
possible) treatment assignments via 3-net clustering randomization following
the algorithm of Ugander et al. (2013), or via unit-level randomization under
a Bernoulli(p) distribution. Sampling of treatment vectors is without
replacement.
}
\examples{
# Create adjacency matrix:

adj_matrix <- make_adj_matrix(N = 50, model = 'small_world',
                              seed = 357)

# Assign treatment via 3-net clustering randomization:
make_tr_vec_bernoulli(adj_matrix, p = 0.2, R = 1,
                      cluster = 'yes', seed = 357)

# Assign treatment via unit-level randomization
# under Bernoulli(p) distribution:

make_tr_vec_bernoulli(adj_matrix, p = 0.2, R = 1,
                            cluster = 'no', seed = 357)

}
\references{
Ugander, J. et al. (2013). \href{https://doi.org/10.1145/2487575.2487695}{Graph cluster randomization: Network exposure to multiple universes}.
\emph{Proceedings of the 19th ACM SIGKDD international conference on Knowledge
discovery and data mining}. 329--337.
}
