% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tr_vec_permutation.R
\name{make_tr_vec_permutation}
\alias{make_tr_vec_permutation}
\title{Randomly assign treatment with uniform probability.}
\usage{
make_tr_vec_permutation(N, p, R, seed = NULL, allow_repetitions = FALSE)
}
\arguments{
\item{N}{number of units.}

\item{p}{proportion of units assigned to treatment.}

\item{R}{number of repetitions (treatment permutations). If \code{allow_repetions = FALSE} and \code{R} is bigger than the number of possible treatment assignments,
then \code{R} is truncated to the number of possible treatment assignements.}

\item{seed}{random number for result replicability.}

\item{allow_repetitions}{logical; if \code{TRUE} sampling is with replacement
rather than without replacement.}
}
\value{
An \code{R} \eqn{*} \code{N} numeric matrix. Each row cooresponds to a treatment
assignment vector.
}
\description{
Create a permutation matrix from possible treatment random assignments with
uniform probability.
}
\details{
\code{make_tr_vec_permutation} produces all possible (or a random sample of
all possible) treatment assignments when sampling without replacement, or a
random sample of possible treatment assignments when sampling with
replacement.
}
\examples{
make_tr_vec_permutation(N = 10, p = 0.2, R = 1, seed = 357)
make_tr_vec_permutation(N = 10, p = 0.2, R = 50, seed = 357)
make_tr_vec_permutation(N = 10, p = 0.2, R = 50, seed = 357,
                        allow_repetitions = TRUE)
}
