% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tr_vec_permutation_hierarchical.R
\name{make_tr_vec_permutation_hierarchical}
\alias{make_tr_vec_permutation_hierarchical}
\title{Randomly assign treatment with uniform probability under a two-stage
hierarchical design.}
\usage{
make_tr_vec_permutation_hierarchical(group, c, k, R, seed = NULL)
}
\arguments{
\item{group}{integer vector of length \code{N} of values specifying the \code{N} units'
group membership (or group labels).}

\item{c}{proportion of groups assigned to the high saturation (psi) in the
first stage.}

\item{k}{numeric vector of length 2 of proportions of individuals whithin
groups assigned to treatment in the second stage, given the first stage
treatment low and high saturations (phi and psi), respectively.}

\item{R}{number of repetitions (treatment permutations).}

\item{seed}{random number for result replicability.}
}
\value{
list of \code{R} data frames each of \code{N} observations and 3 variables:
\describe{ \item{\code{group}:}{Integer with group label.}
\item{\code{group_tr}:}{Numeric indicator of group assignment to high saturation
(psi) in the first stage.} \item{\code{indiv_tr}:}{Numeric indicator of
individual assignment to treatment in the second stage.} }
}
\description{
Create possible treatment random assignments with uniform probability under a
two-stage hierarchical design.
}
\details{
\code{make_tr_vec_permutation_hierarchical} produces all possible (or a random
sample of all possible) treatment assignments under a two-stage hierarchical
design where groups are first randomly assigned to a high or a low level of
treatment saturation (psi, phi), and then units within groups are randomly
assigned to treatment with probability equal to their group saturation rate.
Sampling of treatment vectors is without replacement.
}
\examples{
# Simulate a vector of group membership of 30 units
# equally divided into 6 groups:

group <- rep(1:6, each = 30/6)

# Define proportion of groups assigned to
# saturation psi in the first stage:

c <- 1/2

# Define proportion of individuals whithin groups assigned
# to treatment in the second stage, given the assigned
# first-stage low and high (phi, psi) saturation rates:

k <- c(2/5, 3/5)

make_tr_vec_permutation_hierarchical(group, c, k, R = 1, seed = 357)
make_tr_vec_permutation_hierarchical(group, c, k, R = 5, seed = 357)

}
\references{
Hudgens, M.G. & Halloran M.E. (2008). \href{https://doi.org/10.1198/016214508000000292}{Toward causal inference with interference}. \emph{Journal of
the American Statistical Association}, 103(482), 832--842.

Aronow, P.M. et al. (2020). \href{https://arxiv.org/abs/2001.05444}{Spillover effects in experimental data}. \emph{arXiv preprint}, arXiv:2001.05444.
}
