% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarEstPS_function.R
\name{VarEstPS}
\alias{VarEstPS}
\title{Variance of the population average potential outcome for a correctly
specified propensity score model.}
\usage{
VarEstPS(dta, yhat_group, yhat_pop = NULL, neigh_ind = NULL, phi_hat,
  cov_cols, var_true = NULL, trt_name = NULL, scores = NULL)
}
\arguments{
\item{dta}{The data set including (at least) the treatment and covaratiates.}

\item{yhat_group}{An array including the group average potential outcome
estimates where the dimensions correspond to group, individual treatment and
value of alpha.}

\item{yhat_pop}{A matrix including the population average potential outcome
estimates where dimensions are individual treatment and alpha. If left NULL,
it will be calculated using the means of yhat_group.}

\item{neigh_ind}{A list including one element for each neighborhood. That
element includes the indices of the observations in dta that belong to each
neighborhood. Can be left NULL if dta includes a column neigh.}

\item{phi_hat}{A list with two elements. The first one is a vector of
coefficients of the propensity score, and the second one is the random
effect variance.}

\item{cov_cols}{The indeces including the covariates of the propensity score
model.}

\item{trt_name}{The name of the treatment column. If it is 'A', you can
leave NULL.}
}
\description{
Variance of the population average potential outcome for a correctly
specified propensity score model.
}
