% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarEstPS_IE_function.R
\name{VarEstPS_IE}
\alias{VarEstPS_IE}
\title{Variance of the population average potential outcome for a correctly
specified propensity score model.}
\usage{
VarEstPS_IE(dta, yhat_group, yhat_pop, neigh_ind, phi_hat, cov_cols,
  var_true = NULL, trt_name = NULL, scores = NULL, use = "everything")
}
\description{
Variance of the population average potential outcome for a correctly
specified propensity score model.
}
