% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcNumerator_function.R
\name{CalcNumerator}
\alias{CalcNumerator}
\title{Calculate the numerator in the estimator Yhat.}
\usage{
CalcNumerator(Ai_j, Xi_j, coef_hat, alpha, lower = -10, upper = 10,
  re_alpha = NULL)
}
\arguments{
\item{Ai_j}{A vector of length n_i - 1 including the treatment assignment of all but
the unit we are currently considering.}

\item{Xi_j}{The matrix of covariates for all units in the cluster but the one. This
can be the observed values or a counterfactual set of covariates.}

\item{coef_hat}{A vector of the ps coefficients starting with the intercept.}

\item{alpha}{The average probability of treatment among the n_i - 1 units.}

\item{lower}{The lower end of the values for bi we will look at. Defaults to - 10.}

\item{upper}{The upper end of the values for bi we will look at. Defaults to 10.}

\item{re_alpha}{The fixed effect bi that gives average propensity of treatment in
the group equal to alpha. If set to NULL, it will be calculated.}
}
\description{
Calculate the numerator in the estimator Yhat.
}
