% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcNumerator_function.R
\name{CalcNumerator}
\alias{CalcNumerator}
\title{Calculate the numerator in the estimator Yhat.}
\usage{
CalcNumerator(Ai_j, Xi_j, gamma_numer, alpha, re_alpha)
}
\arguments{
\item{Ai_j}{A vector of length n_i - 1 including the treatment assignment of
all but the unit we are currently considering.}

\item{Xi_j}{The matrix of covariates for all units in the cluster but the
one. This can be the observed values or a counterfactual set of covariates.}

\item{gamma_numer}{A vector of the numerator coefficients starting with the
intercept.}

\item{alpha}{The average probability of treatment among the n_i - 1 units.}

\item{re_alpha}{The fixed effect bi that gives average propensity of
treatment in the group equal to alpha.}
}
\description{
Calculate the numerator in the estimator Yhat.
}
