% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IE_function.R
\name{IE}
\alias{IE}
\title{Indirect effect estimates and asymptotic variance.}
\usage{
IE(ygroup, boots = NULL, ps = c("true", "estimated"), scores = NULL)
}
\arguments{
\item{ygroup}{An matrix including the group average potential outcome
estimates where rows correspond to group, and columns to values of alpha.
#' @param boots The results of BootVar() function including estimates of the
potential outcomes from the bootstrap samples.}

\item{ps}{String. Can take values 'true', or 'estimated' for known or
estimated propensity score. Defaults to 'true'.}

\item{scores}{A matrix with rows corresponding to the parameters of the
propensity score model and columns for groups. Includes the score of the
propensity score evaluated for the variables of each group. Can be left
NULL for ps = 'true'.}
}
\description{
Indirect effect estimates and asymptotic variance.
}
