% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarEstPS_function.R
\name{VarEstPS}
\alias{VarEstPS}
\title{Variance of the population average potential outcome for a correctly
specified propensity score model.}
\usage{
VarEstPS(dta, yhat_group, yhat_pop, neigh_ind, phi_hat, cov_cols,
  var_true = NULL, trt_name = NULL, scores = NULL)
}
\description{
Variance of the population average potential outcome for a correctly
specified propensity score model.
}
